[feat] 6주차 과제1 : ViewModel로 reqres API 활용하기
