+++ app/src/main/java/org/android/go/sopt/presentation/home/view/HomeFragment.kt
@@ -6,12 +6,16 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.Observer
+import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import org.android.go.sopt.databinding.FragmentHomeBinding
 import org.android.go.sopt.data.api.ServicePool.friendService
 import org.android.go.sopt.data.response.FriendData
 import org.android.go.sopt.data.response.ResponseFriendDto
 import org.android.go.sopt.presentation.home.adapter.FriendAdapter
+import org.android.go.sopt.presentation.home.viewmodel.FriendViewModel
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -21,7 +25,9 @@ import retrofit2.Response
 class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
     private val binding: FragmentHomeBinding
-        get() = requireNotNull(_binding){}
+        get() = requireNotNull(_binding) {}
+
+    private val viewModel: FriendViewModel by viewModels()
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -33,11 +39,17 @@ class HomeFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        setData()
+
+        /*서버에서 친구 정보를 가져온다*/
+        viewModel.loadFriendData()
+        viewModel.getAll().observe(viewLifecycleOwner, Observer { data ->
+            /*관찰자 안에다가 데이터 변경이 일어날 때 마다 하고 싶은 작업을 적어준다*/
+            setAdapter(data)
+        })
     }
 
-    fun setAdapter(itemList: List<FriendData?>?){
-        Log.e("setAdapter","어댑터 연결 성공")
+    private fun setAdapter(itemList: List<FriendData?>?) {
+        Log.e("setAdapter", "어댑터 연결 성공")
         val friendAdapter = FriendAdapter()
 
         binding.rvFriendList.adapter = friendAdapter
@@ -46,27 +58,6 @@ class HomeFragment : Fragment() {
         friendAdapter.submitList(itemList)
     }
 
-    private fun setData(){
-        friendService.getFriend().enqueue(object: Callback<ResponseFriendDto>{
-            override fun onResponse(
-                call: Call<ResponseFriendDto>,
-                response: Response<ResponseFriendDto>
-            ) {
-                if(response.isSuccessful){
-                    val itemList : List<FriendData>? = response.body()?.data
-                    setAdapter(itemList)
-                }
-                else{
-                    Log.e("error", response.errorBody().toString())
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseFriendDto>, t: Throwable) {
-                Log.e("error", t.message.toString())
-            }
-
-        })
-    }
 
     override fun onDestroyView() {
         super.onDestroyView()
